+++ src/components/Banner/helpers/index.ts
@@ -1,0 +1,4 @@
+export { BannerInfoSvg } from './BannerInfoSvg';
+export { BannerSuccessSvg } from './BannerSuccessSvg';
+export { BannerWarningSvg } from './BannerWarningSvg';
+export { BannerErrorSvg } from './BannerErrorSvg';